Rename index html to index template
- Rename as above.
- Create new index file.
- Copy from template.
- Delete all body contents except script tag.

A new approach is needed because it's too time consuming to retro-fit Bootstrap with flexbox in this case.

Side-by-side comparison with the template index html file will allow me to build the new index html file, replacing Bootstrap with flexbox along the way (but populating with existing CSS and image assets).